backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai==0.27.8
pydantic
python-multipart
